tests: move mouseMove from test_combobox.py to utils
